use a single abstraction for elements
