+++ src/editor/Overlay/AddingAtom.tsx
